Made FishSpawnAreas not-static + more
The weightedTable on FishSpawnAreas has now been updated to accommodate for neighbourChains, and in order to do this, I also did a To-do: "make FishSpawnAreas non-static", which included me updating prefabs that require FishSpawnAreas to be there as a component. (Boss & FishSpawnManager)

Here's hoping it works + that I haven't caused too many issues!