Updates to own scene + testing
Forgot to mention that the previous push also included some changes to FishChild / Fish that should be uncommented in final build.

This is due to testing in my own scenes without wanting / being able to import XR controls.

I'll go through this before 9am, but for now, I think this'll work.